entity rework
- Added protected bool _is_obstacle and public is_obstacle() to
MapObject, to specify if tile with that object on it is passable.
- Removed has_texture from MapObject in favor of std::optional-driven
texture (draw method works completely fine with it).
- Refactored MapObject's initializers to have less copypasted code.
Added arg for initializing _is_obstacle as required argument.
- Removed MapObject initializer which allowed to create object with
same collision events for both types of creatures, since it hasn't been
used anywhere. May re-add in future, if there will be need.
- Purged object type enums and removed unnecessary categories - since
now, there are just two categories, structure (may later rename it to
Building, if this name will be confused with struct) and creature.
- Reworked initializing of GameMap's internal player_id and exit_id to
work with new entity types and properties.
- Reworked is_tile_blocked() to return status based on is_obstacle() of
tile's content (if at least one item in tile returns true - then it
also returns true, else false).